refactor: secondary button border color
